dup prim before calling it
as an improvement over the previous HVM4, pattern-matching isn't done on
C prims; instead, we do them using native λ-matches, on prelude
functions that wrap around the @@prim, like:

@add = λ{λx.λ{λy.@@add(x,y)}}

this greatly reduces complexity of the C prim files, since they're
guaranteed to receive the terms they need, and not sups/eras. yet, this
alonw isn't enough. consider:

    //(λ{λx.λ{λy.@@eq(x,y)}} &A{0,0} 0)
    //---------------------------------
    //! F &A = λ{λx.λ{λy.@@eq(x,y)}}
    //(&A{(F₀ 0),(F₁ 0)} 0)
    //---------------------------------
    //! F &A = λx.λ{λy.@@eq(x,y)}
    //(&A{(λ{F₀} 0),(λ{F₁} 0)} 0)
    //---------------------------------
    //! F &A = λ{λy.@@eq(&A{x0,x1},y)}
    //(&A{(λ{λx0.F₀} 0),(λ{λx1.F₁} 0)} 0)
    //---------------------------------
    //! F &A = λ{λy.@@eq(&A{0,0},y)}
    //(&A{F₀,F₁} 0)
    //---------------------------------
    //! F &A = λy.@@eq(&A{0,0},y)
    //(&A{λ{F₀},λ{F₁}} 0)
    //---------------------------------
    //! F &A = @@eq(&A{0,0},&A{y0,y1})
    //(&A{λ{λy0.F₀},λ{λy1.F₁}} 0)
    //---------------------------------
    //! F &A = @@eq(&A{0,0},&A{y0,y1})
    //(&A{λ{λy0.F₀},λ{λy1.F₁}} 0)
    //---------------------------------
    //! F &A = @@eq(&A{0,0},&A{y0,y1})
    //&A{(λ{λy0.F₀} 0),(λ{λy1.F₁} 0)}
    //---------------------------------
    //! F &A = @@eq(&A{0,0},&A{y0,y1})
    //&A{(λ{λy0.F₀} 0),(λ{λy1.F₁} 0)}
    //---------------------------------
    //! F &A = @@eq(&A{0,0},&A{0,0})
    //&A{F₀,F₁}

note how we still end up with sups on the @@eq call. the reason is that
wnf will reduce it *before* the dup takes place, so, these sups will be
passed to the C prim function. to avoid that, we block prim_call of a
pri that is surrounded by a dup, and implement dup-pri interaction. this
will force @@eq to be dup'd *before* prim_call is called, as needed.